requirements.txt
async-timeout==4.0.2
bidict==0.22.0
click==8.0.4
Deprecated==1.2.13
dnspython==2.2.1
eventlet==0.33.2
Flask==2.0.3
Flask-SocketIO==5.1.1
Flask-WTF==1.0.0
greenlet==2.0.1
async-generator==1.10
attrs==22.1.0
bcrypt==4.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cors==1.0.1
cryptography==38.0.4
exceptiongroup==1.0.4
filelock==3.8.2
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Testing==0.8.1
future==0.18.2
gevent==22.10.2
greenlet==2.0.1
gunicorn==20.1.0
h11==0.14.0
idna==3.4
importlib-metadata==5.1.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
outcome==1.2.0
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
python-engineio==4.3.1
python-socketio==5.5.2
six==1.16.0
typing_extensions==4.1.1
Werkzeug==2.0.3
wrapt==1.14.0
WTForms==3.0.1
attrs==22.1.0
colorama==0.4.6
iniconfig==1.1.1
pluggy==1.0.0
pytest==7.2.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.44
tldextract==3.4.0
tomli==2.0.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.13
Werkzeug==2.2.2
wsproto==1.2.0
zipp==3.11.0
zope.event==4.5.0
zope.interface==5.5.2
